ML Model
The model I just committed to the repo assess user's inputs and outputs a diagnosis.  Please refer to the dataset to gain an idea of the kind of symptoms the user should be able to select from in this case.  I will promptly commit another piece of code that outputs two related diagnoses.  The model has trouble assessing injuries, but a 95% accuracy in diagnosing diseases.